# Remove TokenStream._tokenToIndexMap and replace with binary search