@heejaechang 2018-10-29
@sharwell what is performance benetfit of readonly struct?