@sharwell 2018-10-29
> @sharwell what is performance benetfit of readonly struct?

The struct was effectively immutable but the API didn't mark it as such. The benefit of marking it falls between no change and some improvement.