@heejaechang 2018-10-29
@sharwell sure. +1 of marking the struck as readonly. was just curious whether it also bring in any perf benefit like marking things as sealed in class.