@heejaechang 2018-10-29
looks like we are marking stuff as readonly struct in our code base. so might be nice if we start to use ref/in as well to remove struct copy overhead?