@heejaechang 2018-10-29
reading some readonly struct article on the web. looks like due to copy overhead, using readonly without ref/in feature could actually make perf worse.